+++ src/trimerge-object.test.ts
@@ -6,7 +6,7 @@ import {
 } from './trimerge';
 import { Path } from './path';
 import { CannotMerge } from './cannot-merge';
-import { trimergeJsObject } from './trimerge-js-object';
+import { trimergeObject } from './trimerge-object';
 
 function mockPathTrackingMerger(paths: Path[]): MergeFn {
   return (_orig, _left, _right, path = []): typeof CannotMerge => {
@@ -24,7 +24,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toBe(s2);
     expect(paths).toEqual([[], ['hello'], ['world'], ['there']]);
@@ -39,7 +39,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toEqual({
       hello: 1,
@@ -57,7 +57,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toEqual({
       hello: 1,
@@ -74,7 +74,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toBe(s3);
     expect(paths).toEqual([[], ['hello'], ['world']]);
@@ -89,7 +89,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toBe(s2);
     expect(paths).toEqual([[], ['hello'], ['world']]);
@@ -104,7 +104,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toEqual({ hello: 1, world: 3, there: 2 });
     expect(paths).toEqual([[], ['hello'], ['world'], ['there']]);
@@ -117,7 +117,7 @@ describe('trimergeJsObject', () => {
     const merger = combineMergers(
       mockPathTrackingMerger(paths),
       trimergeEquality,
-      trimergeJsObject,
+      trimergeObject,
     );
     expect(merger(s1, s2, s3)).toEqual({ hello: 1, there: 2 });
     expect(paths).toEqual([[], ['hello'], ['world'], ['there']]);
@@ -126,28 +126,28 @@ describe('trimergeJsObject', () => {
     const s1 = false;
     const s2 = {};
     const s3 = {};
-    const merger = combineMergers(trimergeJsObject);
+    const merger = combineMergers(trimergeObject);
     expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
   });
   it('does not merge if not all objects 2', () => {
     const s1 = {};
     const s2 = false;
     const s3 = {};
-    const merger = combineMergers(trimergeJsObject);
+    const merger = combineMergers(trimergeObject);
     expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
   });
   it('does not merge if not all objects 3', () => {
     const s1 = {};
     const s2 = {};
     const s3 = false;
-    const merger = combineMergers(trimergeJsObject);
+    const merger = combineMergers(trimergeObject);
     expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
   });
   it('does not merge if none are objects', () => {
     const s1 = false;
     const s2 = false;
     const s3 = false;
-    const merger = combineMergers(trimergeJsObject);
+    const merger = combineMergers(trimergeObject);
     expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
   });
   it('does not class instances', () => {
@@ -155,7 +155,7 @@ describe('trimergeJsObject', () => {
     const s1 = new Foo();
     const s2 = new Foo();
     const s3 = new Foo();
-    const merger = combineMergers(trimergeJsObject);
+    const merger = combineMergers(trimergeObject);
     expect(() => merger(s1, s2, s3)).toThrowError(CannotMergeError);
   });
 });
